loadColumnOptions in parent component
