+++ knowage-vue/src/modules/documentExecution/registry/tables/RegistryPivotDatatable.vue
@@ -1,10 +1,13 @@
 <template>
-    <KnPivotTable :id="id" :columns="filteredColumns" :rows="tempRows" :propConfiguration="propConfiguration" :entity="entity" :pagination="pagination" @rowChanged="onRowChanged" @pageChanged="onPageChange"></KnPivotTable>
+    <KnPivotTable :id="id" :columns="filteredColumns" :rows="tempRows" :propConfiguration="propConfiguration" :entity="entity" :pagination="pagination" :comboColumnOptions="comboColumnOptions" @rowChanged="onRowChanged" @pageChanged="onPageChange" @dropdownOpened="addColumnOptions"></KnPivotTable>
 </template>
 
 <script lang="ts">
 import { defineComponent } from 'vue'
+import axios from 'axios'
 import KnPivotTable from '@/components/UI/KnPivotTable.vue'
+// import mockColumns from './mockColumns.json'
+// import mockRows from './mockRows.json'
 
 export default defineComponent({
     name: 'registry-pivot-datatable',
@@ -23,6 +26,7 @@ export default defineComponent({
             filteredColumns: [] as any[],
             tempRows: [] as any[],
             pagination: {} as any,
+            comboColumnOptions: [] as any[],
             lazy: false
         }
     },
@@ -51,6 +55,9 @@ export default defineComponent({
     methods: {
         getFilteredColumns() {
             this.filteredColumns = this.columns
+
+            // MOCK
+            // this.filteredColumns = mockColumns
         },
         loadRows() {
             this.tempRows = this.rows
@@ -59,6 +66,9 @@ export default defineComponent({
                 this.lazy = false
                 this.tempRows = this.tempRows.slice(0, 15)
             }
+
+            // MOCK
+            // this.tempRows = mockRows
         },
         onRowChanged(row: any) {
             this.$emit('rowChanged', row)
@@ -78,6 +88,33 @@ export default defineComponent({
             }
 
             // console.log('TEMP ROWS AFTER CHANGE: ', this.tempRows)
+        },
+        addColumnOptions(payload: any) {
+            console.log('PAYLOAD: ', payload)
+            const column = payload.column
+            const row = payload.row
+            if (!this.comboColumnOptions[column.field]) {
+                this.comboColumnOptions[column.field] = []
+            }
+
+            if (!this.comboColumnOptions[column.field][row[column.dependences]?.data]) {
+                this.loadColumnOptions(column, row)
+            }
+        },
+        async loadColumnOptions(column: any, row: any) {
+            const subEntity = column.subEntity ? '::' + column.subEntity + '(' + column.foreignKey + ')' : ''
+
+            const entityId = this.entity + subEntity + ':' + column.field
+            const entityOrder = this.entity + subEntity + ':' + (column.orderBy ?? column.field)
+
+            const postData = new URLSearchParams({ ENTITY_ID: entityId, QUERY_TYPE: 'standard', ORDER_ENTITY: entityOrder, ORDER_TYPE: 'asc', QUERY_ROOT_ENTITY: 'true' })
+            if (column.dependences && row && row[column.dependences].data) {
+                postData.append('DEPENDENCES', this.entity + subEntity + ':' + column.dependences + '=' + row[column.dependences].data)
+            }
+            await axios
+                .post(`/knowageqbeengine/servlet/AdapterHTTP?ACTION_NAME=GET_FILTER_VALUES_ACTION&SBI_EXECUTION_ID=${this.id}`, postData, { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } })
+                .then((response) => (this.comboColumnOptions[column.field][row[column.dependences]?.data] = response.data.rows))
+            console.log('COLUMN OPTIONS AXIOS: ', this.comboColumnOptions)
         }
     }
 })
