+++ knowage-vue/src/modules/documentExecution/registry/tables/mockColumns.json
@@ -1,0 +1,241 @@
+[
+    {
+        "columnInfo": {
+            "dataIndex": "column_1",
+            "defaultValue": null,
+            "format": "#,###",
+            "header": "store_id",
+            "multiValue": false,
+            "name": "column_1",
+            "precision": 0,
+            "scale": 0,
+            "type": "int"
+        },
+        "editorType": "TEXT",
+        "field": "store_id",
+        "infoColumn": false,
+        "isEditable": false,
+        "isVisible": false,
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_7",
+            "defaultValue": null,
+            "header": "sales_country",
+            "multiValue": false,
+            "name": "column_7",
+            "type": "string"
+        },
+        "editorType": "COMBO",
+        "field": "sales_country",
+        "foreignKey": "rel_region_id_in_region",
+        "grouping": true,
+        "hasDependencies": [
+            {
+                "columnInfo": {
+                    "dataIndex": "column_8",
+                    "defaultValue": null,
+                    "header": "sales_state_province",
+                    "multiValue": false,
+                    "name": "column_8",
+                    "type": "string"
+                },
+                "dependences": "sales_country",
+                "editorType": "COMBO",
+                "field": "sales_state_province",
+                "foreignKey": "rel_region_id_in_region",
+                "hasDependencies": [
+                    {
+                        "columnInfo": {
+                            "dataIndex": "column_9",
+                            "defaultValue": null,
+                            "header": "sales_city",
+                            "multiValue": false,
+                            "name": "column_9",
+                            "type": "string"
+                        },
+                        "dependences": "sales_state_province",
+                        "editorType": "COMBO",
+                        "field": "sales_city",
+                        "foreignKey": "rel_region_id_in_region",
+                        "infoColumn": false,
+                        "isEditable": true,
+                        "isVisible": true,
+                        "subEntity": "rel_region_id_in_region",
+                        "title": "Foreign key - city",
+                        "unsigned": false
+                    }
+                ],
+                "infoColumn": false,
+                "isEditable": true,
+                "isVisible": true,
+                "subEntity": "rel_region_id_in_region",
+                "title": "Foreign key - state",
+                "unsigned": false
+            }
+        ],
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "subEntity": "rel_region_id_in_region",
+        "title": "Foreign key - country",
+        "type": "merge",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_2",
+            "defaultValue": null,
+            "header": "store_name",
+            "multiValue": false,
+            "name": "column_2",
+            "type": "string"
+        },
+        "editorType": "TEXT",
+        "field": "store_name",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "sorter": "ASC",
+        "title": "store name",
+		"type": "merge",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_3",
+            "defaultValue": null,
+            "header": "store_type",
+            "multiValue": false,
+            "name": "column_3",
+            "type": "string"
+        },
+        "editorType": "COMBO",
+        "field": "store_type",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "title": "store type",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_4",
+            "defaultValue": null,
+            "header": "florist",
+            "multiValue": false,
+            "name": "column_4",
+            "type": "boolean"
+        },
+        "editorType": "TEXT",
+        "field": "florist",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "title": "has florist",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_5",
+            "dateFormat": "d/m/Y H:i:s.uuu",
+            "dateFormatJava": "dd/MM/yyyy HH:mm:ss.SSS",
+            "defaultValue": null,
+            "header": "first_opened_date",
+            "multiValue": false,
+            "name": "column_5",
+            "subtype": "timestamp",
+            "type": "date"
+        },
+        "editorType": "TEXT",
+        "field": "first_opened_date",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "title": "first opened date",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_6",
+            "defaultValue": null,
+            "format": "#,###",
+            "header": "store_sqft",
+            "multiValue": false,
+            "name": "column_6",
+            "precision": 0,
+            "scale": 0,
+            "type": "int"
+        },
+        "editorType": "TEXT",
+        "field": "store_sqft",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "title": "store square foot",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_8",
+            "defaultValue": null,
+            "header": "sales_state_province",
+            "multiValue": false,
+            "name": "column_8",
+            "type": "string"
+        },
+        "dependences": "sales_country",
+        "editorType": "COMBO",
+        "field": "sales_state_province",
+        "foreignKey": "rel_region_id_in_region",
+        "hasDependencies": [
+            {
+                "columnInfo": {
+                    "dataIndex": "column_9",
+                    "defaultValue": null,
+                    "header": "sales_city",
+                    "multiValue": false,
+                    "name": "column_9",
+                    "type": "string"
+                },
+                "dependences": "sales_state_province",
+                "editorType": "COMBO",
+                "field": "sales_city",
+                "foreignKey": "rel_region_id_in_region",
+                "infoColumn": false,
+                "isEditable": true,
+                "isVisible": true,
+                "subEntity": "rel_region_id_in_region",
+                "title": "Foreign key - city",
+                "unsigned": false
+            }
+        ],
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "subEntity": "rel_region_id_in_region",
+        "title": "Foreign key - state",
+        "unsigned": false
+    },
+    {
+        "columnInfo": {
+            "dataIndex": "column_9",
+            "defaultValue": null,
+            "header": "sales_city",
+            "multiValue": false,
+            "name": "column_9",
+            "type": "string"
+        },
+        "dependences": "sales_state_province",
+        "editorType": "COMBO",
+        "field": "sales_city",
+        "foreignKey": "rel_region_id_in_region",
+        "infoColumn": false,
+        "isEditable": true,
+        "isVisible": true,
+        "subEntity": "rel_region_id_in_region",
+        "title": "Foreign key - city",
+        "unsigned": false
+    }
+]
